refactor(ai-settings-model-picker): extract shared build-model-catalog transform into src
Move the pure models.dev transform (zod boundary validation + chat/tool filter
+ sanity checks) and the source URL/attribution constants out of the ingest
script into a shared src module, so the release-pre-step ingest and the runtime
catalog refresh apply identical filtering and validation. The bin script
becomes a thin fetch->transform->write wrapper; the pure-transform tests move
next to the shared module (which also brings them into the typecheck program).
