+++ apps/app/bin/vendor-model-catalog.ts
@@ -3,133 +3,35 @@
  *   fetch models.dev api.json → boundary-validate → generation-time filter
  *   (chat + tool-call models only) → write a committed, deterministic JSON asset.
  *
- * Runtime and the build pipeline NEVER fetch: they read the committed
- * `model-catalog-data.json` only. Only this script performs network I/O, and
- * only when run explicitly via `pnpm vendor:models` (developer / release
- * pre-step). See design.md "Build / Release → vendor-model-catalog".
+ * The build pipeline NEVER fetches: it reads the committed
+ * `model-catalog-data.json` only. This script performs network I/O only when
+ * run explicitly via `pnpm vendor:models` (developer / release pre-step); the
+ * runtime counterpart is the opt-in refresh service
+ * (src/features/mastra/server/services/ai-sdk-modules/refresh-model-catalog.ts),
+ * which shares the same transform. See design.md "Build / Release →
+ * vendor-model-catalog".
  *
  * cross-platform: uses only Node's built-in `fetch` and `node:fs` — no curl/rm.
+ *
+ * NOTE: run via `pnpm vendor:models` (which registers bin/dev-esm-resolver.mjs)
+ * — the imported src modules use the app's path-alias/extensionless import
+ * convention, which plain `node bin/...` cannot resolve.
  */
 import { writeFileSync } from 'node:fs';
 import { resolve } from 'node:path';
-import { z } from 'zod';
 
 import {
-  CATALOG_PROVIDERS,
-  type CatalogProvider,
-  isSelectableModel,
-  type ModelsDevModel,
-} from '../src/features/mastra/server/services/ai-sdk-modules/chat-model-filter.ts';
-
-// ─── Output contract (this script's own types) ──────────────────────────────
-// Kept local: runtime code reads the emitted JSON via a typed `import ... with
-// { type: 'json' }` (task 3.1), it does NOT import these types, and no runtime
-// `src/` module imports from `bin/`.
-
-/** provider → generation-time-filtered selectable model ids (bare ids, sorted). */
-type ModelCatalog = Record<CatalogProvider, string[]>;
-
-/** The committed file shape: header (attribution / timestamp) separated from data. */
-interface ModelCatalogFile {
-  readonly _source: string;
-  readonly _generatedAt: string;
-  readonly models: ModelCatalog;
-}
-
-// ─── Constants ───────────────────────────────────────────────────────────────
-
-const MODELS_DEV_URL = 'https://models.dev/api.json';
-const SOURCE_ATTRIBUTION = `${MODELS_DEV_URL} (MIT)`;
+  buildModelCatalog,
+  MODELS_DEV_SOURCE_ATTRIBUTION,
+  MODELS_DEV_URL,
+  type ModelCatalog,
+  type ModelCatalogFile,
+} from '../src/features/mastra/server/services/ai-sdk-modules/build-model-catalog.ts';
 
 const OUTPUT_PATH = resolve(
   import.meta.dirname,
   '../resource/model-catalog-data.json',
 );
-
-// ─── Boundary schema (tolerant / passthrough) ────────────────────────────────
-// Validate ONLY the fields the filter reads (`tool_call`, `modalities.output`);
-// pass every other field/provider through so upstream additions never break the
-// ingest. A shape violation (missing/mistyped required field, `models` not a
-// map) throws — models.dev schema drift must fail loudly, not ship a bad catalog.
-
-const modelEntrySchema = z.looseObject({
-  tool_call: z.boolean(),
-  modalities: z.looseObject({
-    output: z.array(z.string()),
-  }),
-});
-
-const providerSchema = z.looseObject({
-  models: z.record(z.string(), modelEntrySchema),
-});
-
-// ─── Pure transform (network-free, deterministic, unit-tested) ───────────────
-
-/**
- * Transform a raw models.dev api.json object into the committed catalog's
- * `models` map. Pure and deterministic: no network, no clock (the wrapper adds
- * `_generatedAt`), same input ⇒ same output.
- *
- * For each target provider it boundary-validates the shape, applies the
- * generation-time chat+tool filter (`isSelectableModel`), collects bare ids,
- * sorts them, and asserts the result is non-empty (Issue 2: never ship a silent
- * empty catalog). Any invalid shape or empty provider throws.
- */
-export const buildModelCatalog = (
-  apiJson: unknown,
-  providers: readonly CatalogProvider[] = CATALOG_PROVIDERS,
-): ModelCatalog => {
-  const root = z.record(z.string(), z.unknown()).parse(apiJson);
-
-  const emptyProviders: string[] = [];
-  const models = {} as ModelCatalog;
-
-  for (const provider of providers) {
-    const rawProvider = root[provider];
-    if (rawProvider == null) {
-      throw new Error(
-        `models.dev api.json is missing the target provider "${provider}"`,
-      );
-    }
-
-    // Boundary validation: throws on schema drift (wrong `models` shape,
-    // missing/mistyped `tool_call` or `modalities.output`).
-    const parsed = providerSchema.parse(rawProvider);
-
-    // The bare model id is the map key (mirrored on `entry.id` in real data);
-    // use the validated key set so we don't depend on the optional `id` field.
-    // `parsed.models` entries carry the two authoritative fields validated
-    // above; pass each to the shared filter (single source of truth).
-    const selectableIds = Object.entries(parsed.models)
-      .filter(([, entry]) => isSelectableModel(entry as ModelsDevModel))
-      .map(([id]) => id);
-
-    // Sort by code point (locale-independent) to keep the emitted file
-    // deterministic across environments. A bare `localeCompare()` collates per the
-    // runner's default ICU locale and would reorder ids such as `gpt-5.1-chat-latest`
-    // under e.g. a Czech locale, breaking the same-input⇒same-output guarantee.
-    const sorted = selectableIds.toSorted();
-
-    if (sorted.length === 0) {
-      emptyProviders.push(provider);
-    }
-
-    models[provider] = sorted;
-  }
-
-  if (emptyProviders.length > 0) {
-    const counts = providers
-      .map((p) => `${p}=${models[p]?.length ?? 0}`)
-      .join(', ');
-    throw new Error(
-      `No selectable (tool_call && text-output) models found for provider(s): ${emptyProviders.join(
-        ', ',
-      )}. Selectable counts: ${counts}. Refusing to write an empty catalog.`,
-    );
-  }
-
-  return models;
-};
 
 // ─── Thin I/O wrapper (only runs when executed as the entry point) ───────────
 
@@ -178,7 +80,7 @@ export const main = async (): Promise<void> => {
   }
 
   const file: ModelCatalogFile = {
-    _source: SOURCE_ATTRIBUTION,
+    _source: MODELS_DEV_SOURCE_ATTRIBUTION,
     _generatedAt: new Date().toISOString(),
     models,
   };
@@ -187,9 +89,9 @@ export const main = async (): Promise<void> => {
   // keys are emitted in CATALOG_PROVIDERS order (buildModelCatalog iterates it).
   writeFileSync(OUTPUT_PATH, `${JSON.stringify(file, null, 2)}\n`, 'utf-8');
 
-  const counts = CATALOG_PROVIDERS.map((p) => `${p}=${models[p].length}`).join(
-    ', ',
-  );
+  const counts = Object.entries(models)
+    .map(([p, ids]) => `${p}=${ids.length}`)
+    .join(', ');
   // biome-ignore lint/suspicious/noConsole: ingest script — stdout summary is expected
   console.log(`[vendor:models] wrote ${OUTPUT_PATH} (${counts})`);
 };
